Trebuchet : default workspace
Change default screen to screen 0 (far left screen) to match
Launcher3. Clean up default layouts for various device types.

Also fix missing layouts for allAppsCling.

Update: Replace Books with Hangouts
Update: Clean and add non-GMS fallbacks
Update: Fix MMS being in sw70-non-telephony
Update: Cast checkitemplacement to long
Update: Layouts and avoid rounding issues for sw600dp

+++ src/com/android/launcher3/LauncherModel.java
@@ -1321,7 +1321,7 @@ public class LauncherModel extends BroadcastReceiver {
                             + item.cellY + ") occupied by "
                             + occupied.get((long) LauncherSettings.Favorites.CONTAINER_HOTSEAT)
                             [(int) item.screenId][0]);
-                    if (occupied.get(LauncherSettings.Favorites.CONTAINER_HOTSEAT)
+                    if (occupied.get((long) LauncherSettings.Favorites.CONTAINER_HOTSEAT)
                             [(int) item.screenId][0].itemType == LauncherSettings.Favorites.ITEM_TYPE_ALLAPPS) {
                         deleteOnItemOverlap.set(true);
                     }
